+++ app/Http/Controllers/User/MessageController.php
@@ -9,6 +9,8 @@ use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\Storage;
 use GirdPlugins\Base\ArticleFunc;
 use GirdPlugins\Base\BaseFunc;
+use GirdPlugins\Base\LogFunc;      //zc
+use GirdPlugins\Base\UserPowerFunc;      //zc
 
 class MessageController extends Controller {
 
@@ -25,7 +27,20 @@ class MessageController extends Controller {
         return view("User.Message.sMessage", $combine);
     }
 
-    public function aMessage(BaseFunc $baseFunc) {
+    public function aMessage(BaseFunc $baseFunc, LogFunc $logFunc, UserPowerFunc $UserPowerFunc) {
+
+        //操作记录
+        $log_array["log_level"] = 0;
+        $log_array["log_title"] = "向base_message表中插入一条记录";
+        $log_array["log_detail"] = "向base_message表中插入一条记录";
+        $log_array["log_date"] = date('Y-m-d H:i:s');
+        $log_array["log_user"] = session("user.user_id");
+
+        if (!$UserPowerFunc->checkUserPower(9)) {                    //权限验证
+            return redirect()->back();  //跳回上一页     
+        }
+
+
         $recv_user = $_POST['message_recv_user'];
         $gets = DB::table('base_user')->get();
         $i = 0;  //用于记录的增加
@@ -38,6 +53,7 @@ class MessageController extends Controller {
                 $data['message_recv_user'] = $get->user_id;     //插入的是id,//获取接收者名字的id
                 $data['message_send_user'] = session("user.user_id");
                 if (DB::table('base_message')->insert($data)) {
+                    $logFunc->insertLog($log_array);    //插入操作记录
                     $baseFunc->setRedirectMessage(true, "数据插入成功", NULL, "/user_sMessage");
                     break;
                 } else {
@@ -53,14 +69,26 @@ class MessageController extends Controller {
         }
     }
 
-    public function dMessage($message_id, BaseFunc $baseFunc) {
-        dump($message_id);
+    public function dMessage($message_id, BaseFunc $baseFunc, LogFunc $logFunc, UserPowerFunc $UserPowerFunc) {
+
+        //操作记录
+        $log_array["log_level"] = 0;
+        $log_array["log_title"] = "删除base_message表一条记录";
+        $log_array["log_detail"] = "删除base_message表一条记录";
+        $log_array["log_date"] = date('Y-m-d H:i:s');
+        $log_array["log_user"] = session("user.user_id");
+
+        if (!$UserPowerFunc->checkUserPower(9)) {                    //权限验证
+            return redirect()->back();  //跳回上一页     
+        }
+
         $userId = session("user.user_id");    //提取用户id
 
         $get = DB::table('base_message')->where('message_send_user', '=', $userId)->where('message_id', '=', $message_id)->delete();
         if ($get == null) {
             $baseFunc->setRedirectMessage(false, "删除失败", NULL, "/user_sMessage");
         } else {
+            $logFunc->insertLog($log_array);    //插入操作记录
             $baseFunc->setRedirectMessage(true, "删除成功", NULL, "/user_sMessage");
         }
     }
